TASK: Extract `NodeTypeConstraintChecks` from `ConstraintChecks` so `TetheredNodeAdjustments` are happy
The `TetheredNodeAdjustments` use the `NodeTypeChangeInternals` which only require a subset of the `ConstraintChecks`. But as we declare to use the full set everything is validated via phpstan and it is found out that the `$commandHandlingDependencies` are not available and thus a few methods are unusable (which is true as before)

To make phpstan happy we split the trait in two and only use whats needed. The
